ADM-879-hide-retry feat: hide retry button
